# docs/swagger.yaml
basePath: /
definitions:
  dto.ErrorResponse:
    properties:
      error:
        type: string
      success:
        type: boolean
    type: object
  dto.LoginResponse:
    properties:
      token:
        type: string
    type: object
  dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      image:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
...
  models.MovieDetails:
    properties:
      banner:
        type: string
      casts:
        items:
          type: string
        type: array
      created_at:
        type: string
      director_name:
        type: string
      duration:
        type: integer
      genres:
        items:
          type: string
        type: array
      id:
        type: integer
      image:
        type: string
      release_date:
        type: string
      synopsis:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  pkg.PageInfo:
    properties:
      currentPage:
        type: integer
      nextPage:
        type: integer
      prevPage:
        type: integer
      totalData:
        type: integer
      totalPage:
        type: integer
    type: object
  pkg.Response:
    properties:
      error: {}
      message:
        type: string
      pageInfo:
        $ref: '#/definitions/pkg.PageInfo'
      result: {}
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
  /auth/login:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Authenticate user and return a JWT token
      parameters:
      - description: User email
        in: formData
        name: email
        required: true
        type: string
      - description: User password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with token
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Invalid email or password
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Login user
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Register a new user with email, password, and role
      parameters:
      - description: User email
        in: formData
        name: email
        required: true
        type: string
      - description: User password (min 8 characters)
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: User successfully registered
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Invalid input or email already registered
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Register a new user
      tags:
      - Authentication
  /movies:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: Get all movies with pagination, sorting, and search
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 5
        description: Items per page
        in: query
        name: limit
        type: integer
      - default: id
        description: Field to sort by
        in: query
        name: sortBy
        type: string
      - default: asc
        description: Order of sorting (asc or desc)
        in: query
        name: order
        type: string
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MovieDetails'
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get all movies
      tags:
      - Movies
  /movies/{id}:
    get:
      consumes:
      - application/json
      description: Get details of a specific movie by its ID
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.MovieDetails'
              type: object
        "400":
          description: Invalid input
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Movie not found
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get movie by ID
      tags:
      - Movies
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"